requirements.txt
# pip install -r requirements.txt

streamlit
langchain
langchain-core
langchain-groq
python-dotenv
